Workaround the issue where the viewport size does not respect contentinset
NOTE: The view port height will be 100% of the webview even when the bottom inset is set.
This will make things invisible when the scroll is disabled. (modal, menu, etc.)
The issue is worked around by letting web content be able to be seen through the setting equivalent to "overflow: visible;".

See the following threads for more information:
https://stackoverflow.com/a/35472603
https://bugzilla.mozilla.org/show_bug.cgi?id=1135814
https://github.com/mozilla-mobile/firefox-ios/pull/462/files